Update article.md
Scott made edits on top of Shruti's . Ready for Shako's final review then posting on Medium under his name.  
Shako, when posting on medium pay special attention to the amount of text that fits in the title and first paragraph, when you look at the article in preview/surf-the-article-list mode. If it gets cut-off in the middle of an important phrase, consider rearranging the the text a bit. I can help if needed. All yours. When published, we'l target Hackernoon for re-publication.